ACA-142 Added exception handler
