+++ src/main/java/uk/gov/hmcts/reform/managecase/api/errorhandling/BadRequestException.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-public class BadRequestException extends ApiException {
+public class BadRequestException extends CaseAssignedUserRoleException {
     public BadRequestException(final String message) {
         super(message);
     }
