+++ src/main/java/uk/gov/hmcts/reform/managecase/api/errorhandling/CaseRoleAccessException.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(code = HttpStatus.FORBIDDEN)
-public class CaseRoleAccessException extends ApiException {
+public class CaseRoleAccessException extends CaseAssignedUserRoleException {
 
     public CaseRoleAccessException(String message) {
         super(message);
