+++ src/main/java/uk/gov/hmcts/reform/managecase/api/errorhandling/ResourceNotFoundException.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends ApiException {
+public class ResourceNotFoundException extends CaseAssignedUserRoleException {
 
     public ResourceNotFoundException(final String message) {
         this(message, null);
