+++ src/main/java/uk/gov/hmcts/reform/managecase/api/errorhandling/RestExceptionHandler.java
@@ -10,6 +10,7 @@ import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
@@ -17,11 +18,9 @@ import uk.gov.hmcts.reform.managecase.api.errorhandling.noc.NoCApiError;
 import uk.gov.hmcts.reform.managecase.api.errorhandling.noc.NoCException;
 import uk.gov.hmcts.reform.managecase.api.errorhandling.noc.NoCValidationError;
 
-import javax.servlet.http.HttpServletRequest;
 import javax.validation.ConstraintViolation;
 import javax.validation.ConstraintViolationException;
 import javax.validation.ValidationException;
-import java.io.Serializable;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.List;
@@ -47,15 +46,12 @@ public class RestExceptionHandler extends ResponseEntityExceptionHandler {
         return toResponseEntity(status, null, errors);
     }
 
-    @ExceptionHandler(ApiException.class)
+    @ExceptionHandler({CaseAssignedUserRoleException.class})
     @ResponseBody
-    public ResponseEntity<HttpError> handleApiException(final HttpServletRequest request,
-                                                        final ApiException exception) {
-        log.error(exception.getMessage(), exception);
-        final HttpError<Serializable> error = new HttpError<>(exception, request);
-        return ResponseEntity
-            .status(error.getStatus())
-            .body(error);
+    public ResponseEntity<Object> handleApiException(final CaseAssignedUserRoleException ex) {
+        log.error("CaseAssignedUserRoles exception:", ex.getMessage(), ex);
+        var responseStatus = ex.getClass().getAnnotation(ResponseStatus.class);
+        return toResponseEntity(getHttpStatus(responseStatus), ex.getLocalizedMessage(), null);
     }
 
     @ExceptionHandler(ConstraintViolationException.class)
@@ -151,5 +147,15 @@ public class RestExceptionHandler extends ResponseEntityExceptionHandler {
         NoCApiError apiError = new NoCApiError(status, message, code, errors == null ? null : List.of(errors)
         );
         return new ResponseEntity<>(apiError, new HttpHeaders(), apiError.getStatus());
+    }
+
+    private HttpStatus getHttpStatus(ResponseStatus responseStatus) {
+        if (!HttpStatus.INTERNAL_SERVER_ERROR.equals(responseStatus.value())) {
+            return responseStatus.value();
+        } else if (!HttpStatus.INTERNAL_SERVER_ERROR.equals(responseStatus.code())) {
+            return responseStatus.code();
+        }
+
+        return null;
     }
 }
